enhance(admin): Extract new search code, use it also on datasets and chart search
