+++ adminSiteClient/DatasetsIndexPage.tsx
@@ -1,19 +1,18 @@
 import * as React from "react"
 import { observer } from "mobx-react"
 import { observable, computed, action, runInAction } from "mobx"
-import fuzzysort from "fuzzysort"
 import * as lodash from "lodash"
 
 import { AdminLayout } from "./AdminLayout"
-import { highlight as fuzzyHighlight } from "../grapher/controls/FuzzySearch"
 import { SearchField, FieldsRow } from "./Forms"
 import { DatasetList, DatasetListItem } from "./DatasetList"
 import { AdminAppContext, AdminAppContextType } from "./AdminAppContext"
-
-interface Searchable {
-    dataset: DatasetListItem
-    term?: Fuzzysort.Prepared
-}
+import {
+    buildSearchWordsFromSearchString,
+    filterFunctionForSearchWords,
+    highlightFunctionForSearchWords,
+    SearchWord,
+} from "../clientUtils/search"
 
 @observer
 export class DatasetsIndexPage extends React.Component {
@@ -24,39 +23,32 @@ export class DatasetsIndexPage extends React.Component {
     @observable maxVisibleRows = 50
     @observable searchInput?: string
 
-    @computed get searchIndex(): Searchable[] {
-        const searchIndex: Searchable[] = []
-        for (const dataset of this.datasets) {
-            searchIndex.push({
-                dataset: dataset,
-                term: fuzzysort.prepare(
-                    dataset.name +
-                        " " +
-                        dataset.tags.map((t) => t.name).join(" ") +
-                        " " +
-                        dataset.namespace +
-                        " " +
-                        dataset.dataEditedByUserName +
-                        " " +
-                        dataset.description
-                ),
-            })
-        }
-
-        return searchIndex
+    @computed get searchWords(): SearchWord[] {
+        const { searchInput } = this
+        return buildSearchWordsFromSearchString(searchInput)
     }
 
-    @computed get datasetsToShow(): DatasetListItem[] {
-        const { searchInput, searchIndex, maxVisibleRows } = this
-        if (searchInput) {
-            const results = fuzzysort.go(searchInput, searchIndex, {
-                limit: 50,
-                key: "term",
-            })
-            return lodash.uniq(results.map((result) => result.obj.dataset))
+    @computed get allDatasetsToShow(): DatasetListItem[] {
+        const { searchWords, datasets, maxVisibleRows } = this
+        if (searchWords.length > 0) {
+            const filterFn = filterFunctionForSearchWords(
+                searchWords,
+                (dataset: DatasetListItem) => [
+                    dataset.name,
+                    ...dataset.tags.map((t) => t.name),
+                    dataset.namespace,
+                    dataset.dataEditedByUserName,
+                    dataset.description,
+                ]
+            )
+            return datasets.filter(filterFn)
         } else {
             return this.datasets.slice(0, maxVisibleRows)
         }
+    }
+
+    @computed get datasetsToShow(): DatasetListItem[] {
+        return this.allDatasetsToShow.slice(0, this.maxVisibleRows)
     }
 
     @computed get namespaces() {
@@ -78,14 +70,7 @@ export class DatasetsIndexPage extends React.Component {
     render() {
         const { datasetsToShow, searchInput, numTotalRows } = this
 
-        const highlight = (text: string) => {
-            if (this.searchInput) {
-                const html =
-                    fuzzyHighlight(fuzzysort.single(this.searchInput, text)) ??
-                    text
-                return <span dangerouslySetInnerHTML={{ __html: html }} />
-            } else return text
-        }
+        const highlight = highlightFunctionForSearchWords(this.searchWords)
 
         return (
             <AdminLayout title="Datasets">
